CRM_Project/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.25.1
celery[redis]==5.2.7
django-phonenumber-field[phonenumberslite]==7.0.2
django-debug-toolbar==3.8.1
factory-boy==3.2.1
psycopg2-binary==2.9.5
django-environ==0.10.0
# django-cleanup==7.0.0
django-filter==23.1
Pillow==9.5.0
